refactor: change Card to ProjectCard
