--- src/components/Card/types.ts
@@ -1,1 +1,0 @@
-export interface ICardProps {}
