+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as Card } from './Card';
+export { default as ProjectCard } from './ProjectCard';
 export { default as MySocials } from './MySocials';
 export { default as Header } from './Header';
 export { default as Footer } from './Footer';
